[!!!][TASK] Make localization wizard independent of colPos
The localization wizard in the page module offers now a global translate
action per page only. With this change, a user isn't anymore able to
translate content on column basis.

This streamlines the localization process as it lowers the risk of
creating broken localization setups (a.k.a "mixed mode") and it will
simplify the upcoming "Change mode" wizard that enables to switch
between "Free" and "Connected" modes.

Resolves: #84877
Releases: master
Change-Id: Ibfd4641c5a8d3622c86b5a8657af00b2b3122503
Reviewed-on: https://review.typo3.org/56813
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>