[BUGFIX] Do not copy icon of former EXT:lang during build
The `grunt build` command copies icons of the icon repository into
different locations. Since EXT:lang does not exist anymore, the icon of
the former extension must not be copied into the now invalid location.

The configuration is removed from the Gruntfile to avoid a re-creation of
the extension directory.

Resolves: #84931
Related: #84680
Releases: master
Change-Id: I8387528ab074fab0a1e8a9bb492e69fa67490dce
Reviewed-on: https://review.typo3.org/56864
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>