[TASK] Cleanup workspaces backend AJAX calls
The AJAX endpoints for the workspaces backend module
on the PHP side still are 1:1 like the good old ExtJS times.

In order to abstract and separate concerns in this area,
and make the AJAX-based endpoints true PSR-7 compliant
requests and responses, some cleanups are necessary.

In this first step, the AbstractHandler is removed,
as some methods can be removed or simplified in the
child classes.

Resolves: #84937
Releases: master
Change-Id: If390d4b51d19059b18f579691371669c27a337d4
Reviewed-on: https://review.typo3.org/56872
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>