[TASK] Extract workspace preview uri generation
All logic regarding generating preview links for workspaces has been
extracted into the "PreviewUriBuilder" method. Previously this was all
scattered between the WorkspaceService and the legacy "PreviewHook"
which was not used as a hook.

Separating this functionality allows to further split up the entry points
(Controllers / Handlers) and clean up concerns.

Resolves: #84940
Releases: master
Change-Id: I648fa12bab29ecb53ae319938d1086f95abfff58
Reviewed-on: https://review.typo3.org/56875
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>