[BUGFIX] Fix relogin with popup
The re-login popup in BE has some issues:
- An empty frameset is loaded within the popup which is not needed
and makes the code harder to understand (parent.window.opener)
- The popup always shows the logout form, although the relogin should
show the login credentials form (because the view did not get the memo)
- Login.html should not use "../../../../" for referencing
the form

The LoginFrameset controller thus is not used (will be deprecated
in a separate patch).

This patch moves the logic into the LoginController.

Resolves: #83430
Releases: master
Change-Id: If872baf26297c8b75b786c2d8881802b05b1e41a
Reviewed-on: https://review.typo3.org/54278
Reviewed-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: Mathias Schreiber <mathias.schreiber@typo3.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>