[BUGFIX] Show all content elements with mod.web_layout.defLangBinding
The option is meant to show translated content elements
alligned with their translation parent element (connected mode).

However this also made content elements diasppear, which were
added individually in a language (free mode).

With this change, elements for a language in free mode
are also shown, when the option is set.

Resolves: #62550
Releases: master, 8.7
Change-Id: I96f754ecf4ddb98f22d0d8c7d7da6001fa774d6b
Reviewed-on: https://review.typo3.org/56932
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>