[BUGFIX] Fix deletion of invalid tasks
If the class of a registered task is missing, e.g. because the related
extension has been removed, the task itself will be unserialized into
an __PHP_Incomplete_Class object. Checking the task execution is an
invalid access in this case and must be avoided for such tasks to be
deleted.

Resolves: #84973
Releases: master, 8.7
Change-Id: Iaec8a5e346a0d6257a942cbead50b214036ed700
Reviewed-on: https://review.typo3.org/56959
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>