[BUGFIX] Improve Performance for Inline Elements
Invisible and not translated records will not be retrieved anymore.

Currently every inline element will be translated even if the default
language is requested. This causes a massive performance leak because not only
the current record is loaded, but every other record with the same pid.

Not expanded inline children will be also be ignored.

Resolves: #82100
Releases: master
Change-Id: I3cbe0039f5d15c921d81fc3e634ef38aa06cd18a
Reviewed-on: https://review.typo3.org/54519
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>