[BUGFIX] Ignore timezone for admin panel simulated time
By default, the simulated time of the admin panel does not take the server
timezone into account. This leads to several issues described in the
Forge ticket.

This fix makes sure that the timezone is removed from the users passed
value from the admin panel.

Since the admin panel was reworked in TYPO3 9, this patch does not apply
to master.

Resolves: #66793
Releases: 8.7, 7.6
Change-Id: Ifd86a5f78c10cec54a9fbc4210fd8a68c2be7d6f
Reviewed-on: https://review.typo3.org/57080
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Sybille Peters <sypets@gmx.de>
Tested-by: Sybille Peters <sypets@gmx.de>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>