[TASK] CharsetConverter: Remove Case Folding
Special Case Folding within CharsetConverter is not used anymore, and can be removed.

The public property $caseFolding used as runtime cache is removed, as it was never possible
to fill it either to have it accessed from the outside.

The SpecialCasing.txt file has no meaning anymore and can be removed.

Resolves: #85118
Releases: master
Change-Id: Ib770fe99a97ac06c4f9c485e960c681ab21c9113
Reviewed-on: https://review.typo3.org/57090
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>