[TASK] Add require() to php capable cache backend
To allow loading php cache files more than once it is
necessary to add a require() method next to requireOnce()
to cache backends. This can be used if cache entries do
not execute one-time-per-process-only code, if they for
instance do not declare classes. This is needed to execute
for instance the functional test suite multiple times in
one process and to still make use of caching.
The cache_core php cache is affected by this: Those calls
are changed to require() instead of requireOnce() to make
them multi-loadable per request.
Note require() is not yet added to the PhpCapableBackendInterface
since that would be breaking, but marked as todo for v10.0

Resolves: #85648
Releases: master
Change-Id: Id711044e9554587dba38f976bd1c5bab23826de2
Reviewed-on: https://review.typo3.org/57686
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>