[BUGFIX] Have correct test class namespaces
phpunit tends to execute test case files even if the
file comes with a broken namespace and can't be autoloaded.

The patch hopefully fixes all test case files that are
positioned in a wrong folder, has an otherwise broken
namespace or is not yet registered as psr-4 require-dev.

Resolves: #85169
Releases: master
Change-Id: Id8da485b14ab65a883dd90e1a11cfd8e74d85780
Reviewed-on: https://review.typo3.org/57132
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>