[BUGFIX] Remove TypoScript dependency of AdminPanelInitiator
The admin panel initialization is done via middleware before the 
TypoScript is fully initialized, therefor it cannot have dependencies
to parsed TypoScript (and it cannot be initialized at a later state 
because of cross-dependencies with TSFE). 

Before #85104 the checks did not include TypoScript - that state has 
been restored and the dependency has been removed.

Resolves: #85193
Releases: master
Change-Id: Ic576b02a68ddc20f52b4b1647b0e38c51e1b7df8
Reviewed-on: https://review.typo3.org/57147
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>