[TASK] Avoid thousands of removeDotsFromTs in TemplatePaths
This patch avoids calling GeneralUtility::removeDotsFromTs
on the entire TS array, instead selectively calling it only on
the sub-array and only if a sub-array was actually resolved.

In my extremely small testing setup this patch avoids 1,100
calls in FE and 6,000 calls in BE (in page module with fluid
previews being rendered).

Change-Id: I7195aa310711d82de4861df8dcb2a9fbda309520
Releases: master, 8.7
Resolves: #85341
Reviewed-on: https://review.typo3.org/57351
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>