[TASK] Deprecate TYPO3\CMS\Core\Database\PdoHelper
The PdoHelper class with its single method importSql() has
been added for the Pdo cache backend which is also still
the only usage of that helper class.
With the migration to doctrine, that class feels misplaced
and obsolete. The code is moved over as protected method
into Pdo cache backend directly and the PdoHelper class
is marked as deprecated.

Resolves: #85394
Releases: master
Change-Id: I4a22bd193caefced3cc601b2c1cf79f12961981d
Reviewed-on: https://review.typo3.org/57379
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>