[!!!][TASK] Drop "documentation" extension
The docs team is working on a new docs.typo3.org rendering
chain that can come with some new long awaited features.
One blocker is the core internal documentation extension
that needs various hard to maintain resources from
docs.typo3.org.

The suggested solution was to drop the backend inline
manual fetching and display and thus obsolete the
extension.

For v8 and below, the 'latest' state of the various
json, xml and zip files the extension relies on, will
be frozen by the docs team to not break the module.

The documentation extension has various flaws, it for
instance tends to download outdated manual versions, the
acceptance within the community is not very high
and people are used to look up latest docs online.

The patch drops ext:documentation from the core to
free up the docs team from maintaining the clumsy
requirements at some point in time entirely.

Resolves: #85398
Releases: master
Change-Id: I4152381c49d3223b2586a742388cdf940aafe983
Reviewed-on: https://review.typo3.org/57385
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>