[TASK] Migrate Extbase Typo3DbBackend to Context API
Due to the introduction of Contexts, it is now possible
for Extbase to use the global context and to modify
it to set up a new instance of "PageRepository"
with just a context initialized. This way,
dependencies to $BE_USER and $TSFE->sys_page
are removed.

In further steps, the method should actually just work
with the context object which should be handed in,
however, some more work regarding language handling
has to be implemented first.

Resolves: #85413
Releases: master
Change-Id: If657d25943d92a69a754a29c22461d43b97c4095
Reviewed-on: https://review.typo3.org/57401
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>