[BUGFIX] Remove unused Doctrine restriction
During the Doctrine DBAL refactoring, an
additional check for $TSFE->sys_page->showHiddenRecords
was introduced.

First of all: This check is not needed
as the restriction container is checking for
the REAL value correctly by using
$TSFE->showHiddenRecords.

Additionally, it checks for an invalid
property on PHP object property, which
is never a good idea.

Resolves: #85446
Releases: master, 8.7
Change-Id: Ib9dd0aa0197ee7d8379110a8067c31eba2ff73e8
Reviewed-on: https://review.typo3.org/57434
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>