Update Deprecation-81852-DeprecatedUsageOfEXTrsaauth.rst
I had a typo3 8.7 website with EXT:rsaauth active. A common problem with that is, that Users cannot login when using very long passwords. https://forge.typo3.org/issues/85321 The recommended approach is to deactivate ext:rsaauth (as long as using https at the same time). I tried that out, but could not login anymore with my old password. I then found out that at the same time and you need to set loginSecurityLevel to 'normal'. for FE and BE. So I think it makes sense to mention it here.