[BUGFIX] Fix orientiation of carets in panels
The styling of the caret was only applied, if it is a child of
`.panel-collapsed`, which is no official Bootstrap markup.

The CSS is now changed to rotate the caret icon with the correct
markup, too. This fixes the issue in the System Maintenance Area
that the caret rotation is broken.

Resolves: #85503
Releases: master
Change-Id: I095b7339c3ae8d86e6642fb1953d3bff2b042bfb
Reviewed-on: https://review.typo3.org/57492
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Jürgen Venne <venne@schaffrath-digital.de>
Tested-by: Jürgen Venne <venne@schaffrath-digital.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>