[BUGFIX] Recreate PackageStates in install tool if missing
v8 has a fallback if PackageStates.php vanished on a running instance
and the install tool is called, a default PackageStates file is
generated that comes with a minimal set of active packages that
allow login to the backend and activating further packages again.

This broke in v9 with the install tool refactoring and is fixed
with the patch: The install tool detects after successful login
if PackageStates.php misses and creates a new one. Note this
operation is not yet messaged to the user. This could be added
later if a true recovery procedure is added to the install tool.
For now, the sole purpose of the patch is to resurrect an instance.

Resolves: #85634
Releases: master
Change-Id: I0e72a9e8b4f828892838921d722c61520179af1d
Reviewed-on: https://review.typo3.org/57690
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Andreas Wolf <andreas.wolf@typo3.org>
Tested-by: Andreas Wolf <andreas.wolf@typo3.org>