[TASK] Add documentation render directories to .gitignore
Add directory created by documentation rendering to .gitignore.

Rendering the docs locally is helpful for checking if it will render
correctly on the server. This will create the rendered files in
typo3/sysext/*/Documentation-GENERATED-temp/.

Adding this to .gitignore will prevent accidentally adding the generated
documentation to a patch.

Resolves: #85725
Releases: master, 8.7
Change-Id: Ibcff2570b961e1f1add519ead49e565c671a28ba
Reviewed-on: https://review.typo3.org/57857
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>