[BUGFIX] Set constraints in `FrontendRestrictionContainer` depending on workspace context
The `FrontendRestrictionContainer` must not respect the hidden flag, if
the request happens in a workspace context. The reason is that for
example pages that are created in a workspace have a hidden LIVE version,
which makes overlaying impossible.

This patch changes the behavior to act the same as
`PageRepository::enableFields()` does.

Resolves: #85742
Releases: master, 8.7
Change-Id: I5f1b12bcd8a63daf7eaa9bc9c065b96c343f6ab6
Reviewed-on: https://review.typo3.org/57822
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Sebastian Michaelsen <michaelsen@t3seo.de>
Tested-by: Sebastian Michaelsen <michaelsen@t3seo.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>