[TASK] Clean up Admin Panel Code
Due to the Admin Panel refactoring, some helper methods
were deprecated, but need a better description.

Also, the $BE_USER->uc property is set to "AdminPanel"
instead of cryptic string "TSFE_adminConfig" since
AdminPanel is not related to TSFE anymore.

Additionally, some helper methods and properties which
are not in use are removed.

Resolves: #85815
Releases: master
Change-Id: I7f875acd0f883690f69e895c814b6ad335d86601
Reviewed-on: https://review.typo3.org/57870
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>