[TASK] Refactor metadata handling in FAL
Meta data of files handled by FAL is fetched, created and updated in
various places, which makes it hard to maintain the current code base.

Albeit the method `_getMetaData()` is marked as internal, it has been
marked as deprecated as well, because the method is widely used in the
TYPO3 extension universe.

For this reason, a MetaDataAspect is introduced that takes care of meta
data handling on a low-level basis.

In the same run, FAL's `Indexer` is now responsible for creating or
updating such meta data records, the `ResourceStorage` now only tells
whether auto-extraction is enabled. The meta data extraction, based on
registered extractors implementing the `ExtractorInterface` interface,
has been moved into a separate service class.

Resolves: #85895
Releases: master
Change-Id: Icb929a6226777dcea3868ee5c083cf13ff5a71f6
Reviewed-on: https://review.typo3.org/57908
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>