[TASK] Add resolved site object to FormEngine
With pseudo sites we now always have a site object in backend.
The site object is needed multiple times in FormEngine, the
upcoming slug handling adds another usage.
The patch adds a FormEngine data provider to resolve the
site object for any given page or record and adds it as 'site'
to the result array.

Resolves: #85927
Releases: master
Change-Id: I7ebbd2e6591b54a54b6a5b56f95161b4ba6e5cee
Reviewed-on: https://review.typo3.org/57990
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>