[TASK] Remove a series of SOBE assignments
The infamous $GLOBALS['SOBE'] object is still used at various
places is the core. This "talk to the hand" object registered
as runtime global to make itself available as "parent object"
is an even worse anti pattern than the also infamous $pObj.

The patch removes a series of "$GLOBALS['SOBE'] = $this" assignments
in some controllers that have no further dependency to this global
anymore for a while, are easy to review and b/w compatible.

Resolves: #85958
Releases: master
Change-Id: Ia961289c344c96eab250fa7f206c677cb65e047c
Reviewed-on: https://review.typo3.org/58018
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>