[TASK] Deprecate PageRepository->getFirstWebPage()
The method PageRepository->getFirstWebPage() is only used when
no "?id" parameter is given, and no rootpage was resolved. As this is the only
use-case, a more generic "getMenu" method can be used, which does the
same except for not "limiting" the query to one result, so there is a minimal
memory penalty when doing so, however due to Pseudo-Site functionality
this drawback is only in use for rare cases.

Resolves: #85971
Releases: master
Change-Id: I42fec4251963a0ab2baa038defde39cf0eac4bbf
Reviewed-on: https://review.typo3.org/58029
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>