[TASK] Deprecate @cli annotation
Back then, the PHPDoc annotation `@cli` was added to indicate
Extbase CommandController commands to be usable on CLI only
instead of also be usable as a scheduler task.

The scheduler implementation will be refactored in TYPO3 10
and the execution of commands from the backend will vanish.
Therefore it will not be necessary any more to define if
commands can only be used on the command line or not.
In the future, all commands will only be executable
from the command line.

Releases: master
Resolves: #85977
Change-Id: I52330517b51c5c747177cab55e34a721a0df5f67
Reviewed-on: https://review.typo3.org/58037
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>