[TASK] Deprecate @internal annotation in extbase commands
The @internal annotation – which is a regular phpDocumentor annotation
that is used to denote that associated structural elements are elements
internal to the application or library, has been misused by extbase to
tell if a command is internal and thus should not be exposed through
help texts, user documentation etc. – is deprecated without any
replacement.

TYPO3 does no longer support the use of the @internal annotation to
influence the behaviour of the code.

Releases: master
Resolves: #85980
Change-Id: I7f3b998b56e270b30e5193d494d35329f5608603
Reviewed-on: https://review.typo3.org/58039
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>