[BUGFIX] Do not fetch type=none fields from db in list module
A type=none TCA field is the only field type that is entirely ignored
by DataHandler. An incoming value for such a field is discarded in
checkValue_SW() method and not persisted to db.
The list and query modules should not try to fetch this field type
from database, and a type=none field should not be selectable in
list module single table view.
If a field exists in db and should just be formatted and displayed
readOnly, type=input with readOnly=true is a better option.

Resolves: #86092
Releases: master
Change-Id: Ie3b8e966de66bfd7820b6a17853d4ad591480502
Reviewed-on: https://review.typo3.org/58133
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>