[TASK] Switch UserStorageCapabilityService to a renderType
Using a TCA type=user with a custom renderType is much more
elegant since the registered element can then fully access
$this->data and has control of node expansion and all the
other goodies that come with own renderType in comparison
to the shabby limited type=user with userFunc.
The patch switches the 'is_public' type=user userFunc to
a proper renderType and fixes the incomplete HTML of this
field along the way.
Note a type=check with own renderType is not possible in
this case since the config has no items array.

Resolves: #86109
Releases: master
Change-Id: I72507a4660fc64e5c2df60768564a788a4c38a05
Reviewed-on: https://review.typo3.org/58141
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>