[BUGFIX] Only unlock records in BE user log off functionality
The database table `sys_lockedrecords` should only be cleared via
the BackendUtility when a Backend user is logging off. Currently, this
is also called for Frontend Users, which actually removes everything
from the currently logged-in backend users with the same uid.

As this is very bad code design on many levels (lockRecords for
unlocking, no context for the authentication user object etc), this
should be encapsulated within the BackendUser object directly anyway.

For further abstractions, this could also be a hook or something else,
to be even cleaner.

Resolves: #86113
Releases: master, 8.7
Change-Id: I44d91064edb6ec9ef4c148e48b67bdf22da38869
Reviewed-on: https://review.typo3.org/58148
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>