[BUGFIX] Use class name to mark upgrade wizards done
To make the registry for done wizards consistent and stay backwards
compatible once all upgrade wizards are converted to use the
new interface, always store the class name of the wizard
instead of its identifier in the registry.

Additionally we now add checks for valid identifiers
to all API methods that receive one.

Releases: master
Resolves: #86165
Change-Id: I402852c90f5a8d3d0a05bacbd961a4bb4cb86d1e
Reviewed-on: https://review.typo3.org/58197
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>